prepare: improve initialization to nondet values
we just called klee_make_symbolic on every alloca. This is fine,
but slicer won't slice away this call if the value is initialized later
(because it does not know klee_make_symbolic and must assume the worst
case). Instead of calling klee_make_symbolic directly, create new
alloca, call klee_make_symbolic on it, load from it and store the value
to the original alloca. Slicer will then delete this store if there is
real initialization later